Add Deleting Event_Data

+++ scripts/ZoneMinder/lib/ZoneMinder/Event.pm
@@ -403,6 +403,11 @@ sub delete {
       $ZoneMinder::Database::dbh->commit() if ! $in_transaction;
       return;
     }
+    ZoneMinder::Database::zmDbDo('DELETE FROM Event_Data WHERE EventId=?', $$event{Id});
+    if ( $ZoneMinder::Database::dbh->errstr() ) {
+      $ZoneMinder::Database::dbh->commit() if ! $in_transaction;
+      return;
+    }
     ZoneMinder::Database::zmDbDo('DELETE FROM Frames WHERE EventId=?', $$event{Id});
     if ( $ZoneMinder::Database::dbh->errstr() ) {
       $ZoneMinder::Database::dbh->commit() if ! $in_transaction;
